// src/types/businessAds.ts
export interface BusinessAds {
    id: number;
    name: string;
    logo: string;
    category: string;
    phone: string;
    email: string;
    website: string;
    address: string;
    established: string;
    owner: string;
    specialOffers: string;
    services: string[];
    description: string;
    detailedDescription: string;
}